# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Amenity {
  id: number;
  slug: string;
  name: string;
  icon: string;
}

export interface City {
  id: number;
  slug: string;
  name: string;
  nameAr: string;
  region: string;
  description: string;
  imageUrl: string;
  hotelCount: number;
}

export interface Hotel {
  id: number;
  name: string;
  slug: string;
  city: string;
  citySlug: string;
  address: string;
  stars: number;
  rating: number;
  reviewCount: number;
  pricePerNight: number;
  currency: string;
  imageUrl: string;
  images: string[];
  amenities: string[];
  isFeatured: boolean;
  description: string;
  latitude?: number;
  longitude?: number;
}

export type CityDetail = City & {
  hotels: Hotel[];
};

export interface Room {
  id: number;
  hotelId: number;
  name: string;
  type: string;
  pricePerNight: number;
  capacity: number;
  bedType: string;
  size: number;
  amenities: string[];
  imageUrl: string;
}

export type ReviewCategories = {
  cleanliness?: number;
  location?: number;
  service?: number;
  value?: number;
};

export interface Review {
  id: number;
  hotelId: number;
  guestName: string;
  guestCountry: string;
  rating: number;
  title?: string;
  comment: string;
  date: string;
  categories?: ReviewCategories;
}

export type HotelDetail = Hotel & {
  rooms: Room[];
  reviews: Review[];
  nearbyAttractions?: string[];
  checkInTime: string;
  checkOutTime: string;
  policies?: string;
};

export interface CreateReviewRequest {
  guestName: string;
  guestCountry: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  title?: string;
  comment: string;
}

export interface HotelListResponse {
  hotels: Hotel[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateBookingRequest {
  hotelId: number;
  roomId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkIn: string;
  checkOut: string;
  /** @minimum 1 */
  adults: number;
  /** @minimum 0 */
  children?: number;
  specialRequests?: string;
}

export interface Booking {
  id: number;
  reference: string;
  hotelId: number;
  roomId: number;
  hotelName: string;
  roomName: string;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkIn: string;
  checkOut: string;
  adults: number;
  children?: number;
  specialRequests?: string;
  totalPrice: number;
  currency: string;
  status: string;
  createdAt: string;
}

export type GetHotelsParams = {
  city?: string;
  minPrice?: number;
  maxPrice?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  stars?: number;
  /**
   * Comma-separated list of amenity slugs
   */
  amenities?: string;
  search?: string;
  page?: number;
  limit?: number;
  sortBy?: GetHotelsSortBy;
};

export type GetHotelsSortBy =
  (typeof GetHotelsSortBy)[keyof typeof GetHotelsSortBy];

export const GetHotelsSortBy = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  rating_desc: "rating_desc",
  name_asc: "name_asc",
} as const;