# lib/api-zod/src/generated/types/amenity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */

export interface Amenity {
  id: number;
  slug: string;
  name: string;
  icon: string;
}