# lib/api-zod/src/generated/types/city.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */

export interface City {
  id: number;
  slug: string;
  name: string;
  nameAr: string;
  region: string;
  description: string;
  imageUrl: string;
  hotelCount: number;
}