# lib/api-zod/src/generated/types/cityDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */
import type { City } from "./city";
import type { Hotel } from "./hotel";

export type CityDetail = City & {
  hotels: Hotel[];
};