# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  hotelId: number;
  roomId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkIn: Date;
  checkOut: Date;
  /** @minimum 1 */
  adults: number;
  /** @minimum 0 */
  children?: number;
  specialRequests?: string;
}