# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  guestName: string;
  guestCountry: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  title?: string;
  comment: string;
}