# lib/api-zod/src/generated/types/hotel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */

export interface Hotel {
  id: number;
  name: string;
  slug: string;
  city: string;
  citySlug: string;
  address: string;
  stars: number;
  rating: number;
  reviewCount: number;
  pricePerNight: number;
  currency: string;
  imageUrl: string;
  images: string[];
  amenities: string[];
  isFeatured: boolean;
  description: string;
  latitude?: number;
  longitude?: number;
}