# lib/api-zod/src/generated/types/hotelDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */
import type { Hotel } from "./hotel";
import type { Review } from "./review";
import type { Room } from "./room";

export type HotelDetail = Hotel & {
  rooms: Room[];
  reviews: Review[];
  nearbyAttractions?: string[];
  checkInTime: string;
  checkOutTime: string;
  policies?: string;
};