# lib/api-zod/src/generated/types/hotelListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */
import type { Hotel } from "./hotel";

export interface HotelListResponse {
  hotels: Hotel[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}