# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewCategories } from "./reviewCategories";

export interface Review {
  id: number;
  hotelId: number;
  guestName: string;
  guestCountry: string;
  rating: number;
  title?: string;
  comment: string;
  date: string;
  categories?: ReviewCategories;
}