# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Hotels API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  hotelId: number;
  name: string;
  type: string;
  pricePerNight: number;
  capacity: number;
  bedType: string;
  size: number;
  amenities: string[];
  imageUrl: string;
}